requirements.txt
tensorflow-gpu=2.2.0
keras
matplotlib
sklearn